tests: transform "migrate" testsuite into "reset database" testsuite
