+++ .gitignore
@@ -14,5 +14,4 @@
 /docker/.makefile
 /.env.local
 /docker-compose.override.yaml
-/tests/Fixture/MigrationTests/Migrations/
 /tests/Fixture/Maker/tmp/
