+++ tests/Fixture/EntityInAnotherSchema/Article.php
@@ -18,7 +18,7 @@ use Zenstruck\Foundry\Tests\Fixture\Model\Base;
 
 /**
  * Create custom "cms" schema ({@see Article}) to ensure "migrate" mode is still working with multiple schemas.
- * Note: this entity is added to mapping only for PostgreSQ, as it is the only supported DBMS which handles multiple schemas.
+ * Note: this entity is added to mapping only for PostgreSQL, as it is the only supported DBMS which handles multiple schemas.
  *
  * @see https://github.com/zenstruck/foundry/issues/618
  */
